third_party/blink/web_tests/wpt_internal/pending_beacon/pending_beacon-sendnow.html

<!DOCTYPE html>
<head>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="./resources/pending_beacon-helper.js"></script>
</head>
<body>
<script>
  const baseUrl = `${location.protocol}//${location.host}`;
  promise_test(async t => {
    const uuid = token();
    const url = `${baseUrl}/wpt_internal/pending_beacon/resources/set_beacon_count.py?uuid=${uuid}`;

    // Create and send a beacon.
    const beacon = new PendingBeacon(url, {method: "GET"});
    beacon.sendNow();
    // Wait for the beacon to have sent.
    await wait(1000);

    const beaconCount = await getBeaconCount(uuid);
    assert_equals(beaconCount, 1);
  }, "sendNow() sends a beacon immediately.");

  promise_test(async t => {
    const uuid = token();
    const url = `${baseUrl}/wpt_internal/pending_beacon/resources/set_beacon_count.py?uuid=${uuid}`;

    // Create and send a beacon.
    const beacon = new PendingBeacon(url, {method: "GET"});
    beacon.sendNow();
    // Wait for the beacon to have sent.
    await wait(1000);
    let beaconCount = await getBeaconCount(uuid);
    assert_equals(beaconCount, 1);

    // Try to send the beacon again, and verify no beacon arrives.
    beacon.sendNow();
    // Wait for the beacon to have sent.
    await wait(1000);
    beaconCount = await getBeaconCount(uuid);
    assert_equals(beaconCount, 1);
  }, "sendNow() doesn't send the same beacon twice.");

  promise_test(async t => {
    const uuid = token();
    const url = `${baseUrl}/wpt_internal/pending_beacon/resources/set_beacon_count.py?uuid=${uuid}`;

    // Create and send 1st beacon.
    const beacon1 = new PendingBeacon(url, {method: "GET"});
    beacon1.sendNow();

    // Create and send 2st beacon.
    const beacon2 = new PendingBeacon(url, {method: "GET"});
    beacon2.sendNow();

    // Wait for beacons to have sent.
    await wait(1000);
    beaconCount = await getBeaconCount(uuid);
    assert_equals(beaconCount, 2);
  }, "sendNow() sends multiple beacons.");
</script>
</body>
</html>
